+++ Package.swift
@@ -25,9 +25,24 @@ let package = Package(
         // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "PostHog",
+            dependencies: [
+                .target(
+                    name: "libwebp",
+                    condition: .when(platforms: [.iOS])
+                ),
+            ],
             path: "PostHog",
             resources: [
                 .copy("Resources/PrivacyInfo.xcprivacy"),
+            ]
+        ),
+        .target(
+            name: "libwebp",
+            path: "vendor",
+            sources: ["libwebp/src", "libwebp/sharpyuv"],
+            publicHeadersPath: "include",
+            cSettings: [
+                .headerSearchPath("libwebp"),
             ]
         ),
         .testTarget(
@@ -39,7 +54,10 @@ let package = Package(
                 "OHHTTPStubs",
                 .product(name: "OHHTTPStubsSwift", package: "OHHTTPStubs"),
             ],
-            path: "PostHogTests"
+            path: "PostHogTests",
+            resources: [
+                .process("Resources"),
+            ]
         ),
     ]
 )