+++ PostHogExample/ContentView.swift
@@ -92,6 +92,7 @@ struct ContentView: View {
     var body: some View {
         NavigationStack {
             List {
+                
                 Section("General") {
                     NavigationLink {
                         ContentView()